+++ src/option.c
@@ -174,4 +174,5 @@ void set_default_options(void)
    opt_set_int(OPT_SERVER_PORT, 8888);
    opt_set_int(OPT_STDERR_LEVEL, DIAG_DEBUG);
    opt_set_int(OPT_CHECK_SYNTHESIS, 0);
+   opt_set_int(OPT_MISSING_WAIT, 1);
 }