+++ NEWS.md
@@ -20,6 +20,8 @@
 - OSVVM functional coverage is reported as part of functional coverage
 - Add `threshold-<value>` option for coverage to set minimal coun a
   coverage bin must reach to be reported as covered
+- Added a warning for potential infinite loops in processes without
+  sensitivity and lacking any wait statements (from @NikLeberg).
 
 ## Version 1.14.0 - 2024-09-22
 - Waiting on implicit `'stable` and `'quiet` signals now works